Rename main circuit to ui, digital import uses circuit instead of ux, copy newstructs.h to core.h
